Experimental fix for evaluating mathematical expressions when assigned to variables
Previously, the following line of code:
```var x = b + 25```
would result in setting variable x to the string of "b + 25".

With this fix, the code shown above will now be evaluated accordingly with the consideration of mathematical expressions.

The fix has not been heavily tested, so it is still considered experimental since it may have broken other functionality of the language. Please report any issues that arise with this latest commit.